Refactor TabNet model structure and introduce TabNetRegressor class
