+++ pytorch_tabnet/tab_models/__init__.py
@@ -9,3 +9,16 @@ This package contains the TabNet model implementations including:
 These models implement the TabNet architecture as described in the paper
 "TabNet: Attentive Interpretable Tabular Learning".
 """
+
+from .multitask import TabNetMultiTaskClassifier
+from .pretraining import TabNetPretrainer
+from .tab_class import TabNetClassifier
+from .tab_reg import MultiTabNetRegressor, TabNetRegressor
+
+__all__ = [
+    "TabNetClassifier",
+    "TabNetRegressor",
+    "MultiTabNetRegressor",
+    "TabNetPretrainer",
+    "TabNetMultiTaskClassifier",
+]
